Codigo desordenado

+++ pruebaSpring/src/main/java/com/nigmacode/apirest/repository/PerfilRepository.java
@@ -2,5 +2,5 @@ package com.nigmacode.apirest.repository;
 
 import com.nigmacode.apirest.entity.Perfil;
 import org.springframework.data.jpa.repository.JpaRepository;
-public interface PerfilRepository extends JpaRepository<Perfil, Long> {
+public interface PerfilRepository extends JpaRepository<Perfil, Integer> {
 }
